Make deep copy of resume data. Move _IterationTraceLogger class and print_diagnostic function to diagnostic.py
